Tophat: transpile JS during the markdown parse phase
This means we don't have to include babel as part of the package build.
This drops the main.js filesize from 8.5mb to 3.8mb
